Made util helper back to inline code
