+++ packages/react-core/src/components/Pagination/examples/PaginationDynamicStickyBottom.tsx
@@ -1,17 +1,43 @@
-import { useState, useRef } from 'react';
-import {
-  Pagination,
-  PaginationVariant,
-  Gallery,
-  GalleryItem,
-  Card,
-  CardBody,
-  useIsStuckFromScrollParent
-} from '@patternfly/react-core';
+import { useLayoutEffect, useState, useRef } from 'react';
+import { Pagination, PaginationVariant, Gallery, GalleryItem, Card, CardBody } from '@patternfly/react-core';
+
+const useIsStuckFromScrollParent = ({
+  shouldTrack,
+  scrollParentRef
+}: {
+  /** Indicates whether to track the scroll top position of the scroll parent element */
+  shouldTrack: boolean;
+  /** Reference to the scroll parent element */
+  scrollParentRef: React.RefObject<any>;
+}): boolean => {
+  const [isStuck, setIsStuck] = useState(false);
+
+  useLayoutEffect(() => {
+    if (!shouldTrack) {
+      setIsStuck(false);
+      return;
+    }
+
+    const scrollElement = scrollParentRef.current;
+    if (!scrollElement) {
+      setIsStuck(false);
+      return;
+    }
+
+    const syncFromScroll = () => {
+      setIsStuck(scrollElement.scrollTop + scrollElement.clientHeight < scrollElement.scrollHeight);
+    };
+    syncFromScroll();
+    scrollElement.addEventListener('scroll', syncFromScroll, { passive: true });
+    return () => scrollElement.removeEventListener('scroll', syncFromScroll);
+  }, [shouldTrack, scrollParentRef]);
+
+  return isStuck;
+};
 
 export const PaginationDynamicStickyBottom: React.FunctionComponent = () => {
   const scrollParentRef = useRef<HTMLDivElement>(null);
-  const isStickyStuck = useIsStuckFromScrollParent({ shouldTrack: true, scrollParentRef, position: 'bottom' });
+  const isStickyStuck = useIsStuckFromScrollParent({ shouldTrack: true, scrollParentRef });
   const [page, setPage] = useState(1);
   const [perPage, setPerPage] = useState(50);
   const itemCount = 523;
